+++ lndgrpc/dev.py
@@ -1,0 +1,7 @@
+from .common import ln, BaseClient
+from .errors import handle_rpc_errors
+from datetime import datetime
+import binascii
+
+class DevRPC(BaseClient):
+    pass
